backend/backend-20260123-dbudyak/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "website-monitor"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp",
    "circuitbreaker>=2.0.0",
    "psycopg[binary,pool]",
    "psycopg2-binary",
    "python-dotenv",
    "yoyo-migrations",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "website_monitor.*"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
